requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
pip==21.2.2
numpy==1.16.4
streamlit==0.52.1
pillow
keras
tensorflow==2.0.0b1
